enrichment-service/requirements.txt
confluent-kafka==2.3.0
orjson==3.9.15
psycopg2-binary==2.9.9
redis==5.0.4
icmplib==3.0.3
geoip2==4.8.0
